requirements.txt
absl-py==0.10.0
chex>=0.0.2
dill>=0.3.3
dm-haiku>=0.0.3
jax==0.2.20
jaxlib>=0.1.58
jaxline>=0.0.1
ml_collections>=0.1
numpy>=1.18.0
tensorflow>=2.3.1
tensorflow-addons>=0.12.0
tensorflow-datasets>=4.1.0
tensorflow-probability>=0.12.1
typing_extensions>=3.7
wrapt>=1.11.2
scikit-image>=0.17.2
git+https://github.com/deepmind/dm_pix@6465d23f01ccff1148da183cf1f46cc0ab69a4cb#egg=dm_pix
optax>=0.0.9
chex==0.0.8
